backend/requirements.txt
psycopg2-binary
flask==2.2.2
requests
flask_cors